+++ Python-packages/covidcast-py/docs/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+v0.0.9, July 26, 2020
+---------------------
+- New feature: :py:func:`covidcast.plot_choropleth` and :py:func:`covidcast.get_geo_df`
+  add mapping and plotting capabilities that can be used with the data returned by
+  :py:func:`covidcast.signal`. See the :ref:`function documentation <plotting-data>` for
+  more details.
+
 v0.0.8, July 26, 2020
 ---------------------